Sketch out Native Trino Ion Reader
This is an initial commit of an Ion reader using the Native Trino APIs.
There is basic deserialization of Integers, Text, Structs and Lists.
There are numerous todos and open questions, primarily if we want to
continue to use the ion-java-path-extraction library, as the Hive
SerDe does. But I think this reflects the overall structure we want
and with the tests, this should be a good base to iterate on.