Hacked up Pruning of Nested Columns
This change is an incredibly hacky PoC of how pruning nested columns
for Hive PageSources _could_ work.

It appears to work for both Ion and JSON, though the testing is light.
There is a lot of coupling in it that I don't like, but it reflects
what I think is a solution that addresses the most important
optimizations while being reasonable to achieve.

Did I mention this was super hacky? It needs heavy work but I'm
 pushing it because I don't want to lose it.