Implement Native Trino Ion PageSource and FileWriter
This is the initial implementation of a Native Trino Ion Hive format.

It does not handle the full range of Trino or Ion values, nor the set
of coercions that we will likely need to implement. But I think it's a
pretty good start.

I chose to avoid building off of the "Line" abstractions. While that
could be done, and more closely mimics how the Ion Hive SerDe works,
doing so is suboptimal, and for reading is actually more complicated.
This is because while Ion is unschema'ed, each value stream has an
evolving "encoding context." In Ion 1.0 that is the "Symbol Table"
which is a form of dictionary encoding for the field names and other
common text values.

The other high-level choice was _not_ to use the
ion-java-path-extraction library that is used in the Hive SerDe.
The main value in that was to allow users to do path navigation in the
SerDe config. But using it to go into the Block abstractions without
a middle layer (as it uses in Hive SerDe) is complex, and I'm
concerned that some data quality issues are effectively silenced.
Given Trino's current capability set with nested and complex
structures I doubt the value proposition. In fact, using SerDe config
for what can be expressed in SQL is actually an anti-pattern.

I'm not crazy on the seperation of concerns for the IonPageSource
vs Decoder: it feels like there's a bit of coupling right now. So
that may need revisiting.