Refactor IonDecoderFactory, handle null TLVs
This change refactors IonDecoderFactory to enable decoders that don't
actually write to block themselves but are parents of other decoders.
It will be used to support dereference pushdown and ion path extraction.

I also took the opportunity to handle correctly typed and untyped null
top-level-values as the hive serde does. A mistyped top-level null is
still an error, though it is not in the hive serde.