Refactor IonDecoderFactory, Handle Dereference Paths
This change refactors IonDecoderFactory to enable decoders that don't
actually write to block themselves but are parents of other decoders.
I used new abstractions and refactored implementation to handle trino
dereference pushdown; with this change we also prune nested structs.

The abstractions are designed to handle the two use-cases from path
extraction we need to support:
1. Capturing a top-level Ion value as the single column of the Table.
2. Extracting values from lists with ordinal indexes.

Subsequent changes will make those work by SerDe Properties as today.

I also took the opportunity to handle correctly typed and untyped null
top-level-values as the hive serde does. A mistyped top-level value,
whether null or not is still an error, though it is not in ion hive.
That too will be addressed in a subsequent change.