Change Handling of Some Mistypings and Oversized Values
There are a few changes related to how mistyped or oversize Ion
Values are Handled:

* Enable Strict or Lax Path Typing

This change adds a SerDe Property to support non-strict path typing.
It defaults to false to mimic the behavior of the ion-hive-serde
which used path extraction for the top-level-values, whether the user
had defined any extractions or not. Without support for pathing
this is effectively a type-check (or not) for TLVs. With support for
extraction the behavior is a little more subtle than that, so I named
the property for how it will behave. The name is also consistent with
other properties.

I also added some tests for nested mistypings and changed the code
to consistently throw TrinoExceptions.

* IonInts are coerced to Decimals (as Ion Hive does)
* Decimals with more whole digits than fit is now an error
  (oddly this is not an error in Ion Hive)

Fairly exhaustive test cases were added for each of the changes.

* Ensure Timestamps are Rounded Consistently

This change makes it so that Ion Timestamps are rounded using
Timestamps.round(). This makes it so that Timestamp down casting works
consistently with the other formats.

I considered using DecodedTimestamp and the TrinoTimestampEncoders
but those didn't cover picos. So this code uses the same pattern they
do and the Timestamps.round(). This code simply ignores anything after
picos.

* Truncate Char and Varchar Columns

With this change we truncate text that is longer than the length of
the Char or Varchar column. Before this change that caused an error.

Note that this is an error with the Hive Serde, but per the public
Athena docs and the behavior of the other Hive formats, truncation
is preferred.