Change Handling of Some Mistypings and Oversized Values (#12)
* Coerce Ints to Decimals, Handle Oversized Numbers

This commit has a few changes to edges in handling of mistyped or
oversized values. Each change either brings this impl in line with
Ion Hive or makes it stricter.

* IonInts are coerced to Decimals (as Ion Hive does)
* Decimals with more whole digits than fit is now an error
  (oddly this is not an error in Ion Hive)
* TrinoExceptions are thrown for a few cases

Fairly exhaustive test cases were added for each of the changes.

* Ensure Timestamps are Rounded Consistently

This change makes it so that Ion Timestamps are rounded using
Timestamps.round(). This makes it so that Timestamp down casting works
consistently with the other formats.

I considered using DecodedTimestamp and the TrinoTimestampEncoders
but those didn't cover picos. So this code uses the same pattern they
do and the Timestamps.round(). This code simply ignores anything after
picos.